refactor: move each component into its' own file
